requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.6.0
astroid==2.14.2
asttokens==2.2.1
attrs==22.2.0
autopep8==2.3.2
av==12.3.0
Babel==2.10.3
beautifulsoup4==4.11.2
bidict==0.23.1
black==25.1.0
blinker==1.5
CacheControl==0.14.2
cachetools==5.5.0
certifi==2022.9.24
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
colorzero==2.0
comm==0.2.2
contourpy==1.3.1
cryptography==38.0.4
ctranslate2==4.5.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
dill==0.3.6
distro==1.8.0
docutils==0.19
executing==2.1.0
faster-whisper==1.1.1
filelock==3.17.0
firebase_admin==6.6.0
Flask==2.2.2
Flask-SocketIO==5.5.1
fonttools==4.55.4
fsspec==2025.2.0
gcloud==0.17.0
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gpiozero==2.0.1
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface_hub==0.28.1
humanfriendly==10.0
idna==3.3
importlib-metadata==6.8.0
ipykernel==6.29.5
ipython==8.31.0
isort==5.6.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jiter==0.8.2
joblib==1.4.2
jsonpointer==2.3
jsonschema==4.10.3
jupyter_client==8.6.3
jupyter_core==5.7.2
jws==0.1.3
kiwisolver==1.4.8
lazy-object-proxy==1.9.0
lgpio==0.2.2.0
libarchive-c==2.9
libevdev==0.5
llvmlite==0.44.0
logilab-common==1.9.8
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
meson==1.0.1
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
mypy==1.0.1
mypy-extensions==0.4.3
nest_asyncio==1.6.0
networkx==3.4.2
noisereduce==3.0.3
numba==0.61.0
numpy==1.24.2
oauth2client==3.0.0
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.20.1
openai==1.59.9
openai-whisper==20240930
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.3
pexpect==4.8.0
picamera2==0.3.25
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
Pillow==9.4.0
piper-phonemize==1.1.0
piper-tts==1.2.0
platformdirs==2.6.0
pocketsphinx==5.0.4
prompt_toolkit==3.0.50
proto-plus==1.25.0
protobuf==5.29.3
psutil==5.9.4
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.11.0
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.14.0
pyinotify==0.9.6
PyJWT==2.6.0
pylint==2.16.2
PyOpenGL==3.1.6
pyOpenSSL==23.0.0
pyparsing==3.2.1
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyrsistent==0.18.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-engineio==4.12.0
python-jwt==2.0.1
python-socketio==5.13.0
pyttsx3==2.98
pytz==2022.7.1
pyudev==0.24.0
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
reportlab==3.6.12
requests==2.32.3
requests-oauthlib==1.3.0
requests-toolbelt==0.10.1
responses==0.18.0
rfc3987==1.3.8
roman==3.3
rpi-lgpio==0.6
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
simple-websocket==1.1.0
simplejpeg==1.6.6
simplejson==3.18.3
six==1.16.0
smbus2==0.4.2
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.3.2
SpeechRecognition==3.14.0
spidev==3.5
ssh-import-id==5.10
stack-data==0.6.3
sympy==1.13.1
thonny==4.1.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.11.7
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.64.1
traitlets==5.14.3
twython==3.8.2
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.108
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==1.26.12
v4l2-python3==0.3.5
wcwidth==0.2.13
webcolors==1.11.1
webencodings==0.5.1
webrtcvad==2.0.10
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
zipp==1.0.0
